tools/python/routing/requirements.txt
configparser
GitPython
